updated ui, organized routes, and other small changes
