+++ app/src/Login.css
@@ -25,3 +25,6 @@
   height: 1px;
   background-color: black;
 }
+.account-form {
+  width: 100%;
+}
